refactor(run): split run.py into dispatcher + runners + postprocess
`semi/run.py` was 580 lines pre-refactor and 502 after the bcs.py
extraction; this commit takes it down to 74 lines (a thin dispatcher
plus `SimulationResult`).

Layout chosen: `runners/` package over a flat layout, since the two
solvers share `_common.reference_material` and the package boundary
makes the responsibility split visible at the import line. Resulting
sizes:

    semi/run.py                  74    SimulationResult + run() dispatcher
    semi/runners/__init__.py     21    re-exports
    semi/runners/_common.py      23    reference_material helper
    semi/runners/equilibrium.py  68    run_equilibrium (Day 1 path)
    semi/runners/bias_sweep.py  294    run_bias_sweep + _resolve_sweep
    semi/postprocess.py         141    fmt_tag, record_iv, evaluate_current_at_contact, resolve_contact_facets

All under the 300-line cap. Public API unchanged: `semi.run.run(cfg)`
still dispatches on `solver.type`. `SimulationResult` stays in
`semi.run` for backward compat. The names `run_equilibrium`,
`run_bias_sweep`, `_fmt_tag`, and `_resolve_sweep` remain importable
from `semi.run` via a `__getattr__` shim so the V&V suite and the
existing pure-Python tests keep working without edits.

No behavioral change. Verified:
  pytest: 139/139 pass
  pn_1d:                V_bi 0.8334 V, peak |E| 104.84 kV/cm (unchanged)
  pn_1d_bias:           J(0.6 V) 1.635e+03 A/m^2 (unchanged)
  pn_1d_bias_reverse:   |J|/SRH-gen worst 16.6% (unchanged)
  V&V suite: 53 PASS / 0 FAIL, finest-pair rates byte-identical to baseline
